Moved FromButton to its own file.
